% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatOptions.R
\docType{data}
\name{.formatOptions}
\alias{.formatOptions}
\alias{resetFormatOptions}
\alias{getFormatOption}
\alias{setFormatOption}
\alias{formatOptions}
\title{Global format options}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.formatOptions

resetFormatOptions()

getFormatOption(name)

setFormatOption(name, value)

formatOptions(...)
}
\arguments{
\item{name}{character string of the value to get or set.}

\item{value}{the value to change x into.}

\item{...}{modify options using name = value.}
}
\description{
Global format options
}
\keyword{datasets}
